Fix Mastodon redirect in prod
When using the "Create an application" API you cannot specify more than one redirect uri (you can but only one is valid, the others are ignored). This means that the dev Mastodon client must be different than the prod Mastodon client and the MastodonClientProvider must be aware of that.

This change accounts for the separation between dev and prod by using the current environment to pull the right secrets from Key Vault. Also fixes the dev inner loop by not requiring me to swap the redirect uri between the local and prod url.