Fix reverse search cross street response
When using the Get Reverse Search Cross Street API sometimes the response is a cross street that is not the closest available cross street. Examples: Eastlake Ave East and Eastlake Ave or Pike Street and East Pike Street.

Instead make the API return the top 3 cross streets and use the distance from the search position to pick the best cross street to use. Testing this has shown to fix the above examples and return a better result.