Cache bike lane data
Every time the website was loaded a call would be made to get all the bike lanes. The bike lane data most likely updates very slowly so now we cache the results for 1 hour. 1 hour was picked as if the data does update the updates will be available to users relatively quickly compared to updating less frequently.