Fix time for submissions made not in Seattle
If someone uploads a photo with a user specified time but their current timezone isn't America/Los_Angeles the time on their upload would be converted to the time in their zone. This change fixes that so the time they select will not get changed so the time will reflect the time of the report in Seattle.