Only create a Bluesky session on Admin page open
Ran into rate limits and creating a new session on uploads. Don't need to make a new session upload. On Admin page open, Bluesky session is created and necessary info is saved to the page. On uploads the access token is used to send requests instead of making a new session.